CDSR-650: move case class from controller to model
